Prevent filter from overflowing container
