+++ packages/domain-search/src/components/search-bar/style.scss
@@ -1,0 +1,3 @@
+.domain-search__search-bar-filters {
+	flex-shrink: 0;
+}
